Feature/eint23 376 extract spire api logic

See merge request intellecteu/blockchain/clients/equilend-integration/eint-1source-integration!39

+++ src/main/java/com/intellecteu/onesource/integration/dto/SettlementDto.java
@@ -17,7 +17,7 @@ public class SettlementDto {
     @JsonProperty("partyRole")
     private PartyRole partyRole;
     @JsonProperty("instructionId")
-    private Integer instructionId;
+    private Long instructionId;
     @JsonProperty("settlementStatus")
     private SettlementStatus settlementStatus;
     @JsonProperty("instruction")
